# Optimize locks and refactor for performance